requirements.txt
feedparser>=6.0.10
httpx>=0.27.0
PyYAML>=6.0
anthropic>=0.39.0
